+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.1.7.0 (beta) - 2025-03-XX
 * Smart exposure selector will now select the exposure with lowest percent complete when multiple plans have equal avoidance scores.
+* Added additional cancellation check in TS Container.
 
 ## 5.1.6.0 (beta) - 2025-03-11
 * Added target acquisition summary report to Reporting display.